+++ cx-portal/src/components/pages/AdminBoard/AdminBoardElements.tsx
@@ -18,7 +18,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { useState } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
 import {
   CardDecision,
   PageNotifications,
@@ -27,28 +27,29 @@ import {
 import { useNavigate } from 'react-router-dom'
 import { useTheme, CircularProgress } from '@mui/material'
 import { useTranslation } from 'react-i18next'
+import { show } from 'features/control/overlay/actions'
 import './AdminBoard.scss'
 import {
   AppContent,
   useApproveRequestMutation,
-  useDeclineRequestMutation,
 } from 'features/adminBoard/adminBoardApiSlice'
+import { OVERLAYS } from 'types/Constants'
+import {
+  currentErrorType,
+  currentSuccessType,
+  setErrorType,
+  setSuccessType,
+} from 'features/adminBoard/slice'
 
-export default function AdminBoardElements({
-  apps,
-  handleApproveDeclineSuccess,
-}: {
-  apps?: AppContent[]
-  handleApproveDeclineSuccess: any
-}) {
+export default function AdminBoardElements({ apps }: { apps?: AppContent[] }) {
+  const dispatch = useDispatch()
   const navigate = useNavigate()
   const theme = useTheme()
   const { t } = useTranslation()
-  const [showSuccessAlert, setShowSuccessAlert] = useState<boolean>(false)
-  const [showErrorAlert, setShowErrorAlert] = useState<boolean>(false)
 
   const [approveRequest] = useApproveRequestMutation()
-  const [declineRequest] = useDeclineRequestMutation()
+  const isDecisionSuccess = useSelector(currentSuccessType)
+  const isDecisionError = useSelector(currentErrorType)
 
   if (apps && apps.length === 0) {
     return (
@@ -64,31 +65,28 @@ export default function AdminBoardElements({
     )
   }
 
-  const handleDecision = async (appId: string, status: string) => {
-    const statusFn =
-      status === 'approve' ? approveRequest(appId) : declineRequest(appId)
-    await statusFn
+  const handleApprove = async (appId: string) => {
+    await approveRequest(appId)
       .unwrap()
       .then(() => {
-        setShowSuccessAlert(true)
-        handleApproveDeclineSuccess(true)
+        dispatch(setSuccessType(true))
       })
-      .catch((error) => setShowErrorAlert(true))
+      .catch((error) => dispatch(setErrorType(true)))
   }
 
   const onAlertClose = () => {
-    setShowSuccessAlert(false)
-    setShowErrorAlert(false)
+    dispatch(setSuccessType(false))
+    dispatch(setErrorType(false))
   }
 
   return (
     <div className="admin-board-elements-main">
       <PageSnackbar
-        open={showSuccessAlert || showErrorAlert}
+        open={isDecisionSuccess || isDecisionError}
         onCloseNotification={onAlertClose}
-        severity={showSuccessAlert ? 'success' : 'error'}
+        severity={isDecisionSuccess ? 'success' : 'error'}
         description={
-          showSuccessAlert
+          isDecisionSuccess
             ? t('content.adminBoard.successMsg')
             : t('content.adminBoard.errorMsg')
         }
@@ -97,8 +95,10 @@ export default function AdminBoardElements({
       {apps && apps.length ? (
         <CardDecision
           items={apps}
-          onDelete={(appId: string) => handleDecision(appId, 'decline')}
-          onApprove={(appId: string) => handleDecision(appId, 'approve')}
+          onDelete={(appId: string) =>
+            dispatch(show(OVERLAYS.DECLINE_ADMINBOARD, appId))
+          }
+          onApprove={(appId: string) => handleApprove(appId)}
           onClick={(appId: string) => navigate(`/adminboarddetail/${appId}`)}
         />
       ) : (